% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_linear_model.R
\name{\%r\%}
\alias{\%r\%}
\title{Fitted linear model}
\usage{
y \%r\% x
}
\arguments{
\item{y}{Corresponds to the output vector (which we try to predict).}

\item{x}{Corresponds to the input vector.}
}
\value{
Coefficients of the fitted linear model.
}
\description{
This function calculates the regression coefficients of a linear model.
}
\examples{
2:1 \%r\% 3:2

## linear regression
cars$speed \%r\% cars$dist
}
