% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person_class.R
\name{Person_class}
\alias{Person_class}
\title{Person_class object}
\usage{
Person_class(name, age)
}
\arguments{
\item{name}{Name of a Person}

\item{age}{Age of a Person}
}
\value{
An object of class `Person_class`
}
\description{
Person_class object
}
\examples{
Person_class(name = "Alice", age = 30)

# example of class usage
Bob <- Person_class(name = 'Bob', age = 25)
class(Bob)
}
