% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_info.Person_class}
\alias{print_info.Person_class}
\title{Print method for a class Person_class}
\usage{
\method{print_info}{Person_class}(x, ...)
}
\arguments{
\item{x}{an object to which we apply our method}

\item{...}{Additional parameters (ignored)}
}
\value{
Prints information about the person
}
\description{
Print method for a class Person_class
}
\examples{
Person_class(name = "Alice", age = 30)

}
