% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Use_method_print_info.R
\name{print_info}
\alias{print_info}
\title{Generic function bar}
\usage{
print_info(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Some additional parameters (ignored)}
}
\value{
Method depends on class
}
\description{
Generic function bar
}
